moved menu scene screen to its own file
